move W_OpArgList from opimpl.py to list.py, to break the import cycle between the two files
