+++ spy/vm/opimpl.py
@@ -291,12 +291,3 @@ class W_OpImpl(W_Object):
 
 
 W_OpImpl.NULL = W_OpImpl(None)  # type: ignore
-
-
-# XXX temporary
-from spy.vm.list import W_List, make_prebuilt
-w_oparglist_type = make_prebuilt(W_OpArg)
-W_OpArgList = Annotated[W_List, w_oparglist_type]
-
-def make_oparg_list(args_wop: list[W_OpArg]) -> W_OpArgList:
-    return w_oparglist_type.pyclass(w_oparglist_type, args_wop)  # type: ignore
